Removed extra files, renamed functions_new.py and run_new.py
